renaming middleware to handler and fixing tests
