--- src/middleware.js
@@ -1,3 +1,0 @@
-import handler from './error-handler';
-
-export default handler;
